requirements.txt
fastapi
uvicorn
pydantic
tinytuya
python-dotenv
requests